feet: register route done
