+++ src/controllers/user.ts
@@ -1,46 +1,39 @@
 import type { FastifyInstance, FastifyPluginOptions, FastifyReply, FastifyRequest } from "fastify";
-import { prisma } from "utils/prisma.js";
-import bcrypt from "bcrypt";
+import type UserModel from "models/user.js";
+import type { UserServiceError_t } from "services/user.js";
 
 
 
+/**
+ * @param request
+ * @param reply 
+ * @returns 
+ */
 export const userRegisterController = async (request: FastifyRequest<{ Body: {name: string, email: string, password: string} }>, reply: FastifyReply) => {
 
-    const { name, email, password } = request.body; // as { email: string, password: string};
-
-    // 1. Check if user exists
-    const isUserExist = await prisma.user.findUnique({ 
-        where: { email }
-        // where: { OR: [{ email }, { name }] },
-    });
-
-    if (isUserExist) {
-        return reply.code(400).send({ error: "Email or username already exists" });
-    }
-
-    // 2. Hash password
-    const hashedPassword = await bcrypt.hash(password, 10);
-
-    // 3. Create user
-    // fastify.service.user.create({ name, email, password } as UserModel);
-    const user = await prisma.user.create({
-        data: {
-            name,
-            email,
-            password: hashedPassword,
+    try {
+        const user: UserModel | null = await request.server.service.user.create({
+            name: request.body.name,
+            email: request.body.email,
+            password: request.body.password
+        });
+        if (user !== null) {
+            reply.code(201).send({
+                message: 'user created!',
+                user: {
+                    id: user.id,
+                    username: user.name,
+                    createdAt: user.createdAt
+                }
+            });
+        } else {
+            reply.code(500).send({ message: 'Tir Bzkok!' });
         }
-    });
-
-    // 4. Generate JWT
-    const token = request.server.jwt.sign({ id: user.id, email: user.email });
-
-    reply.code(201).send({
-        message: '201 Created',
-        user: {
-            id: user.id,
-            username: name,
-            createdAt: user.createdAt
-        },
-        token
-    });
+    } catch (e: UserServiceError_t | unknown) {
+        reply.code((e as UserServiceError_t).code).send({
+            statusCode: (e as UserServiceError_t).code,
+            error: (e as UserServiceError_t).message
+        });
+    }
+    
 }
